Basic removable media changing
Prefixes "cdrm" for CDROMs and "zipd" for Zip drives. It sorts
alphabetically but is compatible with the FlashFloppy filename format
DISKAnnnn so users should be able to name their files
CDRM0001-description1.iso to CDRM9999-description2.iso and
be loaded in that order.

Added ".img" suffix to the list of accepted file formats.