Add ATAPI command "read sub channel"
Implemented Read Sub Channel command by importing code from
the ZuluSCSI firmware and massaging it to fit ZuluIDE firmware.

In method cdromGetAudioPlaybackStatus on line 745 in ide_cdrom.cpp
there is a macro switch for the define ENABLE_AUDIO_OUTPUT with code for
CD audio playback that is unaltered from ZuluSCSI firmware.

It should be changed when audio playback is implemented on the ZuluIDE,
otherwise it can be removed to make the code cleaner.