Load media data on insert
This fixes an issue where on an insert after an eject only
loaded the file for the new image. Now on insert the file is
analyzed for image information. For example if a .bin cdrom image
is loaded, it now gathers the audio track data for audio playback
on a mixed or audio CD. Before it would just have change the underlying
data for the previous image.