Correctly sort images with a prefix
Images with a CDRM, ZIPD, and REMV prefix will now load alphabetically.
It will go to the next image in alphabetical order when ejected and
wrap to the first image when the last one is ejected.

If more than one prefix is found, it will choose one and log if it
found others and is ignoring them.