Remove the command interrupt check completely
Removing the check completely that was causing the issue with
the BeOS CD as it was causing read issue with certain IDE controllers
with other CD images.

Commented the code out but left a `// \todo` comment as it might
be that reading the FPGA status command that is causing some sort of
timing issue.