Fix the device interface not being loaded
The global variable `g_ide_device` was never being set so the ide
interface was always null. Put the switch that sets it back into
the code base.

Also modified a return in system_status.cpp it was the same variable
begin negated and logically && to itself. Made the final return
`bool SystemStatus::LoadedImagesAreEqual` `false`. Which means
the function returns false if both values are null or the pointers
aren't the same.