Disable controller if device is not on the the bus
This checks if the IO expander is on the I2C bus. If it is, it
initializes the controller board. If it doesn't fine the expander
it does not initialize the the controller board and also turns off
polling for input.

The controller board polling code slows down reads by about 0.2MB/s
By disabling the polling when there is no controller board the
reads run at their original speed: ~7.4MB/s