Put controller polling code on second core
This is a simple attempt to poll the controller board from the 2nd core.
There seems to be bus contention in the form of loading code for the
2nd core. Even when the polling is skipped there is still a slow down
from 7.4MB/s to 7.2MB/s. The only way to claw back the difference is
to disable the 2nd core when no controller board is detected.

Even though there isn't any speed up, there is a benefit from running
the controller board polling on the second core as the UI is much more
responsive when the IDE bus is busy.