Remove platform dependent code from src
An attempt to move the platform dependant code out of the src directory
when using two cores of the RP2040. Specifically putting loop1 and
setup1 for the second core into the RP2040 platform library.

Moved the Control MVC into it's own library.

Increased the throughput of the IDE bus by 0.6 MB/s by putting the
the platform and SDFat static libraries into SRAM.