Support Execute Device Diagnostic command
The Phoenix BIOS seems to do the ATA command `Execute Device Diagnostic`
before issuing the `Identify Device` ATA command. This commit
attempts to support the diagnostic command, but is only designed
as a primary device, with or without a secondary device on the bus.

Further code will be needed to support the command when the device
is secondary.

There is a an outstanding issue with Phoenix BIOS support as the device
causes an error in the boot process. After the BIOS sends the
`Execute Device Diagnostic` command it issues a `Identify Packet Device`
when it should be issuing a `Identify Device` command.

This causes the BIOS to pause the boot process waiting for an `F1`
or `F2` on the test machine used. If the boot process is continued
Linux issues the proper `Identify Device` command and is able to
access to drive on the test machine used.