More accurate zip drive emulation  (fpga update)
This disables IORDY which the Zip drive doesn't use. It also returns
a status 0x58 instead of 0x48 during a ATA command which more accurately
mirrors the Zip drive. This is controlled by setting
`g_ide_config.disable_iordy = true` which is currently hard coded for
testing.

It seems like the main issue seen in the waveforms is the random IDE
bus INTRQ assertions during the data transfer of the identify device
ATA command. Which causes the Host (Roland sequencer) to issue commands
it normally wouldn't.