First attempt at configurable zip drive setup
This is a first attempt on getting basic zip drive support setup
via filename and ini file settings. To properly support it, there
will need to a concept of drive type and drive media type.
Currently drive and media types are set via the medium type.

Which is especially noticeably when a SD card is removed and reinserted.
Because IDE drive type is generally queried only once, it should
only be set once. Where the media type should be able to be changed
depending on what is currently loaded.