On insert media report device getting ready
The device reports busy (spinning up) after a media is inserted.
This fixes an issue where FreeDOS does not list the proper directory of
the current CD after the previous one has been ejected.

On ejects the media gets cycled, that means issuing multiple ejects
will cycle the media multiple times allowing the user to skip media.
After the tray closes, the device reports it is spinning up until a
Test Unit Ready is sent which it reports busy one last time and then
acts as if the drive is fully spun up.

Needs testing with other OSes.