Got proper cdrom ejection working under linux
In Lubuntu gui, the drive properly ejects with the button and then
inserts when the eject button is pressed again.

This process uses the `Get Event Notification` request eject. For some
reason Windows is not responding to this and can only eject an image
via the eject feature in Explorer.