Roland SP-808 ejection button working
The button ejection process on the Roland SP-808 now works. The Roland
is able to read that the ejection button was pressed though the Zip
drive 0x06 vendor command to prompt saving the latest song and then
eject the disk.

Also added is the ability to ignore the host blocking removal of media.
This is done with a new `zuluide.ini` setting:
`ignore_prevent_removal = 1`