Add ability to insert disk on SD card insert
The `zuluide.ini` setting `reinsert_media_on_sd_insert = 1` will insert
the media if the device is in the eject state and the SD card has been
reinserted.

The setting is useful for the SP-808 because it allows one less step
when ejecting and inserting media with the ZulueIDE. Also fixed
an issue with the SP-808 thinking media has been inserted while the
SD card is ejected and the ZuluIDE is in the ejected state.