UDMA Writes mostly working - getting CRC errors
For some reason the very last block of a multiple sector write is
getting a CRC error. The difference being a `FPGA_CMD_ATA_READ` is sent
instead of a `FPGA_CMD_ATA_READ_CONT`. Have not done a data integrity
test to see if the data is actually bogus or not.