Add .ini files to the ignore list
Windows can create desktop.ini files in a folder and is normally hidden
from the users unless system files are visible. This change will ignore
those and other .ini files that might be mistaken for images.

It also removes a function the ide_imagefile.cpp that is no longer used,
related to validating an image file.