This fixes an issue on an ASUS TUSI-M
An FPGA change was required along with turning off IORDY was required
to make the ASUS TUSI-M boot with the ZuluIDE CD-ROM detect every time.

The problem was the IDE controller was requesting a PIO data read
without issuing a command causing the ZuluIDE to goto a busy state
without it being cleared by a real PIO data read.