Split return DSC and disable IORDY
Disabling IORDY used to be linked to returning Disk Seek Complete (DSC)
when Disk Ready was set in the status register. This was because of
work on the zip drive and rigid hard drive support.

This change splits the setting into individual settings and lets IORDY
to be set via a ini file setting `disable_iordy`.
It also includes a change to the fpga bit stream.