Split status reg DSC support to phy settings
The ASUS TUSI-M CD-ROM drive support seems to require the DSC be
returned with DRDY when set in the status register.

To make this an ini file setting, DSC support was removed from the
config struct and added to the phy setting command that currently
supports IORDY signaling.

It DSC support is now on by default for the CD-ROM drive emulation
so the ASUS TUSI-M's IDE controller and BIOS work by default.