Add ATA Device Reset support
The fix set the registers properly for the ATA command Device Reset.
It doesn't release the IDE signals it is supposed to when the command is
issued, but it does work to boot the OAKCDROM DOS CD-ROM driver.

The other fix to get the DOS driver to work is returning the error bit
in the ATA status register for signature returning ATA commands.
Specifically when Device Identify is aborted because the device uses the
 ATA Packet command and the proper command for ident is Device Packet
 Identify.