Add feature to set or unset ATAPI INTRQ signaling
It seems some OSes need an INTRQ signal between the ATA Packet command
and the ATAPI CDB command, others need the signal repressed.

A setting has been added to the zuluide.ini file under `[IDE]`
as `atapi_intrq`. When set to 1 the signal is enabled, when set to 0
this signal is disabled.

FreeDOS seems to need it disabled where as Linux seems to needs it
enabled. There might be an underlying issue cause this. But this
is the current work around.