Fix ZuluIDE emitting "Not Ready" on image change
Some OSes expect a "Not Ready" error emitted once after a media change.
It was assumed that the ATAPI command "Test Unit Ready" would be
called afterwards and the the "Not Ready" error could be cleared there.

DOS does not do this, so the clearing of the "Not Ready" was moved
to atapi_cmd_error call.