Fix issue with MS-DOS videcdd CD-ROM driver
The driver would take a few seconds to `dir` the first time.
This issue was resolved by returning the ATA status register with
a Disk Seek Complete (0x10) ORed with the standard Disk Ready (0x40).

Another issue was uncovered once first bug was squashed. An ATAPI
error `Not Ready` was issued one time on the next ATAPI command after a
media change (disc insert). The videcdd driver immediately fails on
the next `dir`. The `Not Ready` functionality is now a `zuluide.ini`
setting under `[IDE]` as `set_not_ready_on_insert` which is false
or 0 by default, but can be set to 1 or true to enable the `Not Ready`
functionality.