Load last image on initialization
When switching between CD-ROM images with the ZuluIDE, a file
`zululast.txt` will be created keeping track of the last image used.
On initialization, ZuluIDE will load the file listed in `zululast.txt`
if it can find it.

Also fixed the issue where changing images by using the OS to eject
a CD would not update the control board display. Now it will.