Fix Zip Drive support
There was a bug in the ide_atapi code where the zip drive wouldn't
change images in Linux and probably other OSes. Added unit attention/
medium change to when images are inserted which fixed the issue.

Image changing via the control board doesn't work correctly with
the Roland SP-808. But the eject button still does. There need to be
a way to queue image changing for the next ejection for when drives
are ejection locked.