Fix multi-bin file audio playback stutter
getTrackFromLBA was parsing the whole cue file whenever it was called.
This caches the current track and the ending LBA of the track. If it
is called during audio playback and the lba is still within the track,
the cached track info is returned.

The cache is clear when set file is called.

Haven't tested the edge cases yet.