Fix audio playback stuttering on the last track
Fixed calculating last track LBA upper limit. The upper limit requires
getting the file size of the last track. The last track was failing to
open the last bin file in the `getTrackFromLBA` method. Switched to the
`selectBinFileForTrack` method and used the capacity method on the
m_image object to get the file size so the upper LBA limit on the last
track could properly be calculated. This allow the cached
`getTrackFromLBA` to be used on playback of the last track.