If Zip drive eject prevented, defer loading image
These changes will defer the loading of an image file if the ATAPI
command Prevent Allow Medium Removal is set to prevent. Changes are
mainly made to how the controller board will handle and display this
status. For example if the eject feature is used, the display
will print `--- Host Eject Only ---`. When an image is selected
the bottom line scrolls `To load image, eject device from host system`
and then the message disappears once the host has ejected the previous
image and loads the next one.