Fix Pico-W controller from crashing the board
With a Pico-W web interface controller, changing an image would
cause the ZuluIDE to crash. The cause seemed to be running out
of SRAM for a new iterator object. The solution was to remove
'LoadImageByFileName' utility methods and reuse the already allocated
iterator in the I2CServer class when changing images.