Fix referenced nullptr in the eject method
The member variable `m_image` can be a nullptr. For example when
an image is ejected via the I2C controller board, this will clear
m_image. If eject method gets run after this, the attempt to get the
filename for the eject message hits a hard fault. This fix checks
m_image before attempting to eject and gives a different eject message
if the image is cleared.