Fix initialization without controller board
When there was no I2C controller board hooked up to the ZuluIDE,
the I2C Pico W webserver was testing it's I2C connection before it set
its I2C interface so it was acting on a nullptr.

Fixed messaging when either the SSID or password is not set. Before
if either one was not set it would report both not being set.