Fix USB mass storage
After moving the initialization code to initVariant the mass storage
was being checked before the USB system was initialized. The fix
was take the arduino-pico framework USB initialization function and
execute before the mass storage device is checked.

Also moved the debug flag to after the the SD card is initialized.