Fix BIOS detection issues caused by status BSY
Award BIOS on VIA chipset appears to do an unexpected read of
the IDE DATA register on boot. ATA specification says that this
should be ignored if STATUS DRQ is not active.

Currently the FPGA code does not ignore this transfer, but instead
sets the STATUS BSY bit to signal that the data buffer is empty.
This caused ZuluIDE access to be stuck until the timeout handler
would reset STATUS register.

This commit makes the STATUS register reset time shorter (10 ms)
and resets it to device ready state. This way the BIOS detection
can proceed immediately. Also changed the log message to a debug
message because the condition itself is harmless.