FPGA: Support splitting one UDMA sector into multiple bursts
VIA 82C686 terminates UDMA transfers in the middle of a sector
when its buffer fills up, and expects the device to continue.
Previously ZuluIDE would hang in this situation.

Now DMARQ is reasserted if more data remains.
FPGA automatically combines the CRC checksums from all bursts
of a single sector.