Split CHS init from main init
The main device initialization was moved after the first image is loaded
for setting the CHS of a rigid disk. Unfortunately the main init
zeros out some setting that the first image loading sets.

This was causing the first CD-ROM loaded unable to play audio because
it's CD type was being zeroed out by the device initialization.

This fix splits out the CHS code into a method called `post_image_setup`
and is called after the first image is loaded. The device initialization
code has been moved back above the loading of the first image.