User friendly SD card errors
Suppress timeout errors that occur when the SD card is removed.
Other SD errors are still logged to the user if they occur.
In debug mode the timeout errors are included in the logging.

Changed the behavior of dbgmsg, it now returns the state of g_log_debug.
This allows printing a different message if debug is disabled that may
be more user friendly. Or in this case, the code filters what should be
included in the normal logmsg while everything is included in the
dbgmsg.