Fix Win9X CD playback and proper TOC reporting
Proper track changes for Win9x cd player. There was a change needed
to the audio_get_lba_position() due to the file_offset in the CUEParser
being the start of index1 in the cue file and the start of the track
being index0 before index1. This was causing a negative number in
a calculation of unsigned numbers.

This should also pulls in a change to the CUEPraser library to give
proper TOC timings if a PREGAP is in the cue file.