RP2040: Replace data direction asserts with debug messages
Previously the phy code would assert() and crash the firmware
if device code queries status of data transfer and the direction
does not match the FPGA status.

This especially occurred if something went wrong with ATAPI packet
commands.

Changed it to just report a debug message. This will at least allow
returning proper error to the host and hopefully recover.