CDROM: Fix infinite loop when ejecting from Linux
Running "eject /dev/cdrom" from Linux resulted in an infinite loop
in two ways.

When reinsertion was disabled:
1. Linux sends start_stop with eject
2. ZuluIDE reports event EjectRequest
3. In response to event, Linux sends another eject request and cycle repeats.

When reinsertion was enabled:
1. Linux sends start_stop with eject
2. ZuluIDE reports event EjectRequest and reinserts media
3. In response to event, Linux sends another eject request and cycle repeats.

It may be that we shouldn't report EjectRequest event when the ejection
was requested through IDE bus and not by button.

For now I fixed this by reporting the event only if there was previously
media present, and delaying the reinsertion by 500 ms.