Decode imageName url parameter string
This is for issue: https://github.com/ZuluIDE/ZuluIDE-firmware/issues/192
Filenames with "&" fail to load via the web interface with a PicoW
control board.

The issue is the filename string was being sent as a url without being
encoded. Since "&" is special character as a URL parameter the
string wasn't being received properly by the ZuluIDE.

Now encoding URL parameter in the https://github.com/ZuluIDE/ZuluIDE-HTTP-PicoW
repo and and escaping the url in this repo for the ZuluIDE firmware.