Fix infinite loading of CDs in Linux
The use of a delay to stop Linux from loading and looping all images one
after another was causing another issue so it was removed. A different
approach is used to stop the infinite loading of CDs, by auto-inserting
CDs on the get event notification and only emitting event notification
for a requested eject when prevent allow removal is respected e.g.
`ignore_prevent_removal = 0` in the `zuluide.ini`.