Hide controller board splash screen after 3 secs
This is to address the issue: https://github.com/ZuluIDE/ZuluIDE-firmware/issues/177

The issue is that if there is no SD card the splash screen on the
controller board doesn't disappear until a SD card is loaded.

The fix is to force a status update, in this case setting the firmware
version, after 3 seconds if the splash screen is still active.