Remove logging messages during filename transfer
This greatly increase the reliability sending names to the Pico Sheild.

Some issue remaining, the update should only occur at boot and
when the SD card presence is changed. Instead it is occurring when an
image is changed. The waiting period between sending files should
probably adjusted as it takes a few seconds for 300+ filesnames to be
transferred.