Adjustable number of tick before a rotary action
The number to ticks of the rotary encoder before the ZuluIDE sends
a clockwise or counter clockwise action is adjustable. It can be set
in the `zuluide.ini` file under [UI] as `ticks = x` where x
is the number of ticks. The default is now 1.

Switched to a state machine to decode the rotary encoder. This has
two benefits over the previous version.
 1. It effectively debounces the inputs by ignoring their state changes
 2. It uses both rising and lower edges of channel A instead of just one