Moved rotary ticks init to when SD is first read
After testing the splash screen when there is no SD card, it was noticed
that the number of ticks it takes to select a menu item with the rotary
switch set in the `zuluide.ini` file wasn't being respected.

This moves the reading the number of ticks from `zuluide.ini` to
the platform poll loop. It only sets the value first time the SD card is
loaded so it reads the `zuluide.ini` file just once for the setting.