Fix CHS reporting zeros and capacity zero size
With certain rigid disk drives the capacity doesn't fit in CHS
combinations. With those drive sizes CHS values will be chosen that are
smaller than the actual capacity but valid.

Reads were failing because the drive capacity was being reported as
zero. This is because the insert call was not implemented and was
nulling the image pointer. The fix has insert call pass the image
pointer to set image.