Fix slow transfer rate of multi bin file cd images
Track information was being properly cached to speed up the transfer
rate of the cue/multi-bin files CD images, but a call to capacity_lba
caused the cache to be invalidated so the on every read the whole
cue file was being processed each time.

The solution is to cache the return value of capacity_lba and
invalidate the capacity value when the cached track information
is invalidated. Now the transfer rate is the same as the other CD
image formats.