I2C server mostly works with image response pipe
Image iteration is now only run on the first core for serving up
images to the I2C server running on the second core.

The main issue now is while image iteration is occurring if there
is SD card access from the IDE interface, the iteration stops before
all the images have been processed. If this occurs, the Wifi I2C
client's drop down list of images to load will not be complete.