Fix SDIO error during I2C server image iterations
ZuluIDE executes platform_poll while setting up reads. This was causing
an SDIO error. Added flag to platform_poll that lets the loop know
if it is executing from the ZuluIDE.cpp main loop and not during a read.
Running the I2C Server update for image iteration only from the this
loop avoids the SDIO error.