Get both I2C Server and Control Board working
Was able to handle the controller board and the I2C Server code
on the first core while handling IDE requests.

It may make sense to have the I2C transfers to go over DMA for better
response times from the reset of the system. Currently while the
I2C is polling for a transfer to finish it runs ide_protocol_poll()
until the transfer is done, thus starving other subsystems.