Clean up log messaging
Parsing of images that is only done when an SD card is initially
found give warnings as to what images were ignored and their reasons.
This can be turned off by setting the zuluide.ini setting of
`quiet_image_parsing` to 1.

The iterator no longer logs a message if it could not open the root
directory.

Prevent the loading of the image twice on boot by implementing
the image.cpp ""=="" method by comparing filenames, previously it
just returned false.