Ignore hidden directories and files
The image_iterator was looking at hidden files and folders. Now it
silently ignores them.

Directories where being validated with the same filename criteria as
files. Now directories are no longer filtered with the filename
criteria.

Directories with the first character non-alphanumeric were being checked
for cue files. Now only directories with the first character that is
alphanumeric are being checked for cue files and there is a warning
for the first character being non-alphanumeric.