Implement IOCS16 signaling for PIO data transfers
This is a simple implementation of IOCS16 signaling for ZuluIDE V2s.
The ZuluIDE V2 as a CD-ROM drive now works on Digital Celibris with a
CMD PCI-0640 PCI to IDE controller. If this change breaks compatibility
with other IDE controllers, it can be turned off with the `zuluide.ini`
file setting `disable_iocs16 = 1` under the `[IDE]` section.