Disable auto UDMA mode by default
This commit moves turning on UDMA mode automatically into a zuluide.ini
setting and disables it by default. This is to address older machines
that do not follow the ATA/ATAPI spec fully. For example the CMD
PCI-0640 PCI to IDE controller running Windows 98. Sometimes
the feature register doesn't get cleared to 0x00 before a Packet command
is issued, causing the ZuluIDE to assume it is a UDMA transfer because
the first bit of feature register is still set to 1.

The general use case for UDMA mode being set automatically is when
the device unexpectedly reset due to a bus error or watch dog timeout.

This also contains a CORE 1 change that fixes a separate issued
for the PCI-0640 controller.