Fix stack overflow with multiple image files
The firmware was crashing with a stack overflow when multiple images
had to be iterated through.

Merged temporary filename strings into a single temporary string
off the stack in `ide_cdrom.cpp` and made temporary filenames in the
image iterator static so they to would not be place on the stack.

Tested with over 300 iso images and was crash free.

Also excluded "bootloader.elf" from the `analyze_crashlog.sh` script
as it would usually be found before firmware.elf and not display
the proper stack trace.