Add initial support for large CUE files
An issue customers encountered was their CUE files were to large
for the ZuluIDE to handle. This could not be easily fixed by increasing
the CUE buffer sizes. Up to 4 x 4k buffers for cue files corresponding
to different instances of the CUE parser could be allocated at one time.

The following solution uses one shared CUE buffer.

To do this a sub class around the CUEParser library called the
SharedCUEParser is used. It uses a shared buffer between objects to
store the CUE file. With each public method of CUEParser it first checks
to see if the CUE file currently in the shared buffer is for the current
object. If it isn't, it reads the correct CUE file into the buffer from
the SD card.

This is initial support without much error checking.