Fix stack overflow with shared cue parser
After implementing a large shared buffer for CUE files a stack overflow
error was detected on the V2. It may have also been happening on the
RP2040 since it doesn't have stack overflow protection.

To fix the issue a set method for the SharedCUEParser was created.
This is so loadAndValidateCueSheet function can avoid creating a new
object and then copying the object to member SharedCUEParser object.
Now the set method modifies the object directly without the need
to allocated anything.

Also set a filename string to static in loadAndValidateCueSheet to
save space on the stack.