Fix stack overflow when zululast.txt isn't present
For multi bin/cue files when zululast.txt file is not preset a stack
overflow occurs. This fix created a temp filepath static class array,
a static temp FsFile handler along with making the FsFile handler for
the root directory of the SD card static also. The temporary static
variables are used in place of some of the local variables in member
methods. This saved enough memory on the stack to fix the issue.

Printing out the rejected files and directories no longer does it twice
when no zululast.txt file is present.