Fix dos eject and audio playback on seek
A third party eject command was going into an infinite loop.
The fix was the set the drive to be lockable in the page code 0x2A
"CD Capabilities and Mechanical Status Page" from the SCSI MMC x3
document.

The other fix which occurred after an eject command, was seek always
attempting to play audio at the new position even if audio wasn't
already playing. Causing seeks in data tracks to throw an error message.
Setting the audio to playback at the seek location only if audio is
already in playing stop the error from occurring.