Longer full path and less cue sheet hits
This was originally to address issue: https://github.com/ZuluIDE/ZuluIDE-firmware/issues/259
when concatenating cue sheet filenames.

This adds more guard rails around the cue sheet filenames.
It also doubles the length of the full file path for the SharedCUEParser
So now the full path can be 255 characters for the directory and
255 characters for the CUE filename. The CUE sheet filenames and image
filenames have been normalized to 255 characters for a valid filename.
If the filename is 256 characters it considers the filename to have been
truncated and invalidates the image.

The first and last trackinfo are now stored on the ide_cdrom.cpp class
when an image is set. This reduces the need to read the CUE sheet twice
on set_image() for bin/cue files. It also avoid rereading the CUE sheet
on capacity_lba() and atapi_read_disc_information().

Copyright headers have been added to SharedCUEParser.