Fix/reset volume on CD/image change (#265)
* Reset CD audio volume on disc change

This is for issue: https://github.com/ZuluIDE/ZuluIDE-firmware/issues/263
It seems some DOS apps expect their volume to be reset on disc change.
This fix resets the volume whenever a new CUE file is sent to the
audio system.

* Re-enable volume channels on CD change

Along with the volume level being reset. The drive should probably
re-enable the volume channels. When disabled they act as a mute.

---------

Co-authored-by: J. Morio Sakaguchi <morio.earth@gmail.com>