Add create image from ZuluSCSI
This is a port of create image functionality from the ZuluSCSI firmware.
To creat a zeroed image, prefix a filename with "Create", add the file
size "100MB", then the filename of the image to create
"zipd my zip disk.bin".

The emtpy file would look like this: "Create 100MB zipd my zip disk.bin"
and this should create a 100MB file called: "zipd my zip disk.bin"

This occurs when the board is powered on with the SD card or when an SD
card is inserted while the board is already powered on. The first
allowing the ZuluIDE to load the image as it primary image and select
the device type. With the second only creating an new image, while
the device type will have been set on power on.