Add Microsoft's VHD image container compatibility (#269)
* Add Microsoft's VHD image container compatibility

Adds the ability to use VHD images that are fixed (the container
contains the full size of the drive in the data section).

It will detect a valid VHD file automatically even if the extension is
not present. Dynamic or differencing VHD images may get corrupted if
used.

* Add unsupported container image type reporting

ZCFsFile added unsupported image detection. When detected, logging
a an error message to the user in the log file. The error message
look like the following.
```
[1468ms]
[1468ms] ============ ERROR: Unsupported container image type ============
[1469ms] Image is a vhd container but the image type is unsupported.
[1469ms] Please use a container with a fixed size or fully allocated image
[1470ms] =================================================================
[1471ms]
```