Implement ATAPI Mechanism Status command and fixes
This implements the ATAPI Mechanism Status command which is a required
command in the SCSI MMC 3 spec.

While testing the code it was found the ATAPI command Get Event
Notification was returning an error when no media was present when it
should just return a status OK.

A debatable change is a to have the ZuluIDE in the ejected state
when there is no image loaded. This has been applied to when
the SD card is removed and when the SD is card is not present on power
on. The exception is when `no_media_on_sd_insert` or `no_media_on_init`
are set. This is to avoid the image being auto inserted when
the `reinsert_media_*` settings are also set.