WIP: Get Event Status Notification working
After tracing get event status notification with a Windows 8 ATAPI
bus logger. It was determined that a real DVD player on eject would
notify a condition change, then a MediaRemoval media message.
The SCSI/ATAPI MMC spec says MediaRemoval is only for media changers
(devices that hold multiple discs) but this doesn't seem to be that
case in the real world. On insert it issues a condition change and
then a NewMedia media message.

Tested in both Linux and Windows 7, the drive properly updates the disc
name for the drive in the GUI without having to refresh the drive
folder.

Attempted to make the no_media_on_init and no_media_on_sd_insert code
a little more sane. But ended up breaking the no_media_on_sd_insert.

Windows allows ejection of an ejected disc by sending a Start/Stop
ATAPI command. It would be useful if this command inserted the disc
from a no_media_on_* settings and if the device is in the normal ejected
state.