Get Event Status Notification act as a real drive
After tracing get event status notification with a Windows 8 ATAPI
bus logger. It was determined that a real DVD drive on eject would
notify a condition change, then a MediaRemoval media message.
The SCSI/ATAPI MMC spec says MediaRemoval is only for media changers
(devices that hold multiple discs) but this doesn't seem to be the
case. On insert ZuluIDE issues a condition change and then a NewMedia
media message.

Tested in both Linux and Windows 7, the drive properly updates the disc
name for the drive in the GUI without having to refresh the drive
folder.

Cleaned up the no_media_on_init and no_media_on_sd_insert code.